more bugs are becoming apparant

+++ MSOE.MediaComplete.Lib/Library/DataSource/FileSystem.cs
@@ -33,7 +33,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         private static FileSystem _instance;
 
         /// <summary>
-        /// Returns the instance of the filesystem
+        /// Returns the instance of the file system
         /// </summary>
         public static IFileSystem Instance { get { return _instance ?? (_instance = new FileSystem()); } }
 
@@ -118,6 +118,8 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         public bool DirectoryEmpty(DirectoryPath directory)
         {
             if (!Directory.Exists(directory.FullPath)) return true;
+            var dirInfo = new DirectoryInfo(directory.FullPath);
+            if ((dirInfo.Attributes & FileAttributes.ReadOnly) == FileAttributes.ReadOnly) return true;
             var hasDirs = Directory.EnumerateDirectories(directory.FullPath).Any();
             var hasMusic = new DirectoryInfo(directory.FullPath).EnumerateFiles().GetMusicFiles().Any();
             return hasDirs || hasMusic;
@@ -169,13 +171,14 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         }
 
         /// <summary>
-        /// writes a local song object to a filesystem object
+        /// writes a local song object to a file system object
         /// </summary>
         /// <param name="song"></param>
         public void SaveSong(AbstractSong song)
         {
             var lsong = (song as LocalSong);
-            if (!_cachedSongs.ContainsKey(lsong.Id)) throw new ArgumentException("Song does not exist in cache", "song");
+            if (lsong == null || !_cachedSongs.ContainsKey(lsong.Id)) 
+                throw new ArgumentException("Song does not exist in cache", "song");
             var file = TagLib.File.Create(lsong.Path);
 
             foreach (var attribute in Enum.GetValues(typeof(MetaAttribute)).Cast<MetaAttribute>().ToList()
@@ -202,7 +205,8 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// <param name="dest"></param>
         public void MoveFile(LocalSong source, SongPath dest)
         {
-            File.Move(source.Path, dest.FullPath);
+            if (FileExists(source.SongPath))
+                File.Move(source.Path, dest.FullPath);
         }
 
         /// <summary>
@@ -212,7 +216,8 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// <param name="dest"></param>
         public void MoveFile(SongPath source, SongPath dest)
         {
-            File.Move(source.FullPath, dest.FullPath);
+            if(!FileExists(source))
+                File.Move(source.FullPath, dest.FullPath);
         }
 
         /// <summary>
@@ -222,12 +227,15 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         public void DeleteSong(AbstractSong deletedSong)
         {
             var song = (deletedSong as LocalSong);
-            var sourceDir = song.SongPath.Directory;
-            _cachedSongs.Remove(song.Id);
-            if (song.Path.Equals(_cachedFiles[song.Id].FullName) && File.Exists(song.Path))
-                File.Delete(song.Path);
-            _cachedFiles.Remove(song.Id);
-            ScrubEmptyDirectories(sourceDir);
+            if (song != null)
+            {
+                var sourceDir = song.SongPath.Directory;
+                _cachedSongs.Remove(song.Id);
+                if (song.Path.Equals(_cachedFiles[song.Id].FullName) && File.Exists(song.Path))
+                    File.Delete(song.Path);
+                _cachedFiles.Remove(song.Id);
+                ScrubEmptyDirectories(sourceDir);
+            }
         }
 
         /// <summary>
@@ -470,7 +478,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// Helper function for adding new songs to the dictionaries
         /// </summary>
         /// <param name="id">unique ID of the song to be saved</param>
-        /// <param name="file">The Fileinfo object of the file to be saved</param>
+        /// <param name="file">The FileInfo object of the file to be saved</param>
         private void AddFileToCache(string id, FileInfo file)
         {
             var newFile = GetNewLocalSong(id, file);
@@ -484,7 +492,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// the new LocalSong object.
         /// </summary>
         /// <param name="id">unique ID of the file to be saved</param>
-        /// <param name="file">fileinfo object that needs to be saved</param>
+        /// <param name="file">FileInfo object that needs to be saved</param>
         /// <returns></returns>
         private static LocalSong GetNewLocalSong(string id, FileSystemInfo file)
         {
@@ -544,6 +552,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
             {
                 return true;
             }
+            catch (UnauthorizedAccessException){}
             finally
             {
                 if (stream != null)
@@ -558,17 +567,27 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// <param name="song"></param>
         private void UpdateFile(LocalSong song)
         {
-            var tagFile = TaglibFile.Create(song.Path);
+            try
+            {
+                var tagFile = TaglibFile.Create(song.Path);
+                var tag = tagFile.Tag;
+                _cachedSongs[song.Id].Title = tag.Title;
+                _cachedSongs[song.Id].Artists = tag.AlbumArtists;
+                _cachedSongs[song.Id].Album = tag.Album;
+                _cachedSongs[song.Id].Genres = tag.Genres;
+                _cachedSongs[song.Id].Year = tag.Year;
+                _cachedSongs[song.Id].TrackNumber = tag.Track;
+                _cachedSongs[song.Id].SupportingArtists = tag.Performers;
+                // Duration is assumed to be fixed
+            }
+            catch (FileNotFoundException)
+            {
 
-            var tag = tagFile.Tag;
-            _cachedSongs[song.Id].Title = tag.Title;
-            _cachedSongs[song.Id].Artists = tag.AlbumArtists;
-            _cachedSongs[song.Id].Album = tag.Album;
-            _cachedSongs[song.Id].Genres = tag.Genres;
-            _cachedSongs[song.Id].Year = tag.Year;
-            _cachedSongs[song.Id].TrackNumber = tag.Track;
-            _cachedSongs[song.Id].SupportingArtists = tag.Performers;
-            // Duration is assumed to be fixed
+            }
+            catch (UnauthorizedAccessException)
+            {
+                
+            }
         }
         #endregion
 
@@ -583,7 +602,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
     }
 
     /// <summary>
-    /// Interface for the datasource represented by the local file system
+    /// Interface for the data source represented by the local file system
     /// </summary>
     public interface IFileSystem : IDataSource
     {
@@ -598,7 +617,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// Create a folder at a specified location.
         /// Used by Sorter and to initialize music/playlist folders where necessary
         /// </summary>
-        /// <param name="directory">Destination location to create the folder, including foldername</param>
+        /// <param name="directory">Destination location to create the folder, including folder name</param>
         void CreateDirectory(DirectoryPath directory);
         /// <summary>
         /// Verifies if the specified directory exists.
@@ -671,7 +690,7 @@ namespace MSOE.MediaComplete.Lib.Library.DataSource
         /// <returns></returns>
         LocalSong GetSong(SongPath songPath);
         /// <summary>
-        /// Returns a local song object based on a playlist's mediaitem
+        /// Returns a local song object based on a playlist's MediaItem
         /// </summary>
         /// <param name="mediaItem"></param>
         /// <returns></returns>
